requirements.txt
streamlit
torch
numpy
pandas
scikit-learn
plotly
python-dotenv
google-genai